[wandb] Have `train_loop_config` logged as a config.
